refactor: bip38 worker process
