+++ src/renderer/workers/bip38.js
